refactor: Names가 Name을 관리하도록 역할 분배
