+++ src/main/java/domain/NamesCreator.java
@@ -4,29 +4,12 @@ import java.util.Arrays;
 import java.util.List;
 
 public class NamesCreator {
-    private static final int MIN_NAMES_COUNT = 2;
-    private static final int MAX_NAMES_COUNT = 10;
     private static final String SEPARATOR = ",";
 
     public Names create(String names) {
         validateSeparator(names);
         List<String> splitNames = splitName(names);
-        validateDuplicateName(splitNames);
-        validateNameCount(splitNames);
         return new Names(splitNames.stream().map(Name::new).toList());
-    }
-
-    private void validateNameCount(List<String> splitNames) {
-        if (splitNames.size() < MIN_NAMES_COUNT || splitNames.size() > MAX_NAMES_COUNT) {
-            throw new LadderGameException(ExceptionType.INVALID_NAMES_RANGE);
-        }
-    }
-
-    private void validateDuplicateName(List<String> splitNames) {
-        long distinctCount = splitNames.stream().distinct().count();
-        if (distinctCount != splitNames.size()) {
-            throw new LadderGameException(ExceptionType.NOT_ALLOW_DUPLICATE_NAME);
-        }
     }
 
     private void validateSeparator(String names) {
